Split workspace data structures out of small_neutronnova_zk.rs
SmallAbc, PrefixWorkspace, ExtendedPrefixMleEvals, and the prefix/suffix
fold helpers that operate on them move to a child module
(small_neutronnova_workspace.rs), shrinking small_neutronnova_zk.rs to the
prep/prove pipeline and the accumulator NIFS driver. Pure code motion; the
no-custom-macros introspection test now covers both files.
